pyproject.toml
[project]
name = "piboatcore"
version = "0.1.0"
description = "Raspberry Pi boat telemetry service with heartbeat and sensor health reporting."
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
modem = ["pyserial>=3.5"]

[project.scripts]
piboatcore = "pi_boat_core.service:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
